# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AkadiaPay is a school financial platform API for authentication, payment configuration, student billing, payment processing, allocation, and financial ledger.\n\nBusiness process coverage:\n- Authentication and tenant workspace access\n- Payment policy configuration\n- Payment product configuration\n- Student obligation billing\n- Parent/student outstanding bill visibility\n- Payment order processing\n- Payment allocation\n- Financial ledger posting and reporting",
        "title": "AkadiaPay API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "AkadiaPay API Support",
            "email": "support@akadiapay.local"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/payment-orders/{id}/allocate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Distributes one payment order to one or more student obligations using the existing payment policy and no over-allocation validation rules. Allowed roles: SUPER_ADMIN, SCHOOL_ADMIN, TREASURER, PARENT, STUDENT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Allocation"
                ],
                "summary": "Allocate payment order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment allocation payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PaymentAllocationAllocate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SwaggerPaymentAllocationResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payment-orders/{id}/allocations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns allocation details for one payment order, including allocated totals and remaining amount under current finance portal access rules. Allowed roles: SUPER_ADMIN, SCHOOL_ADMIN, TREASURER, PARENT, STUDENT.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment Allocation"
                ],
                "summary": "Get payment allocations by payment order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SwaggerPaymentAllocationResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation error message"
                }
            }
        },
...
        "domain.PaymentAllocationAllocate": {
            "type": "object",
            "required": [
                "allocations"
            ],
            "properties": {
                "allocations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.PaymentAllocationCreate"
                    }
                }
            }
        },
        "domain.PaymentAllocationCreate": {
            "type": "object",
            "required": [
                "student_obligation_id"
            ],
            "properties": {
                "allocated_amount": {
                    "type": "number"
                },
                "student_obligation_id": {
                    "type": "string"
                }
            }
        },
        "domain.PaymentAllocationResponse": {
            "type": "object",
            "properties": {
                "allocated_amount": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "payment_order_id": {
                    "type": "string"
                },
                "student_obligation_id": {
                    "type": "string"
                }
            }
        },
        "domain.PaymentAllocationResult": {
            "type": "object",
            "properties": {
                "allocations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.PaymentAllocationResponse"
                    }
                },
                "order_status": {
                    "$ref": "#/definitions/model.PaymentOrderStatus"
                },
                "payment_order_id": {
                    "type": "string"
                },
                "remaining_amount": {
                    "type": "number"
                },
                "total_allocated": {
                    "type": "number"
                }
            }
        },
...
        "domain.SwaggerPaymentAllocationResultResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.PaymentAllocationResult"
                }
            }
        },
...
        "model.PaymentOrderStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "COMPLETED",
                "CANCELLED",
                "EXPIRED"
            ],
            "x-enum-varnames": [
                "PaymentOrderStatusPending",
                "PaymentOrderStatusCompleted",
                "PaymentOrderStatusCancelled",
                "PaymentOrderStatusExpired"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Example: \"Bearer eyJhbGciOiJIUzI1NiIs...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}